Create docker-compose.yml
This file defines configuration to run both the api application and the neo4j database in a single container. If the api fails (due to db connection error or other), it will restart. Data and Log resilience is achieved through the "volumes" flag, currently set the /home/neo4j/data and /home/neo4j/log. Note that inside the container, neo4j runs as neo4j user and hence will change permissions on the mapped folders.